粗略完成

+++ src/main/java/com/example/zhihudemo/service/CommentService.java
@@ -24,7 +24,7 @@ public interface CommentService {
 
     int add(Comment comment);
 
-    List<Comment> findByRedId(int refId);
+    List<Comment> findByRedId(long refId);
 
-    List<Comment> query(int refId);
+    List<Comment> query(long refId);
 }
